Move chain ID and height incrementer
