+++ crates/chain/chain-components/src/traits/types/height.rs
@@ -4,6 +4,7 @@
 
 use cgp::prelude::*;
 
+pub use hermes_chain_type_components::traits::fields::height::*;
 pub use hermes_chain_type_components::traits::types::height::*;
 
 #[derive_component(HeightFieldComponent, HeightFieldGetter<Chain>)]
@@ -11,11 +12,6 @@ pub trait HasHeightFields: HasHeightType {
     fn revision_number(height: &Self::Height) -> u64;
 
     fn revision_height(height: &Self::Height) -> u64;
-}
-
-#[derive_component(HeightIncrementerComponent, HeightIncrementer<Chain>)]
-pub trait CanIncrementHeight: HasHeightType + HasErrorType {
-    fn increment_height(height: &Self::Height) -> Result<Self::Height, Self::Error>;
 }
 
 #[derive_component(GenesisHeightGetterComponent, GenesisHeightGetter<Chain>)]
