+++ crates/chain/chain-type-components/src/traits/fields/mod.rs
@@ -1,0 +1,2 @@
+pub mod chain_id;
+pub mod height;
