+++ crates/chain/chain-type-components/src/traits/mod.rs
@@ -1,1 +1,2 @@
+pub mod fields;
 pub mod types;
