+++ crates/chain/chain-type-components/src/traits/types/mod.rs
@@ -1,3 +1,4 @@
+pub mod chain_id;
 pub mod commitment_prefix;
 pub mod commitment_proof;
 pub mod event;
