+++ crates/cosmos/cosmos-chain-components/Cargo.toml
@@ -15,6 +15,7 @@ description  = """
 cgp                                     = { workspace = true }
 hermes-relayer-components               = { workspace = true }
 hermes-relayer-components-extra         = { workspace = true }
+hermes-chain-type-components            = { workspace = true }
 hermes-encoding-components              = { workspace = true }
 hermes-protobuf-encoding-components     = { workspace = true }
 hermes-cosmos-encoding-components       = { workspace = true }
