+++ crates/relayer/relayer-components-extra/Cargo.toml
@@ -12,7 +12,8 @@ description  = """
 """
 
 [dependencies]
-cgp                         = { workspace = true }
-hermes-relayer-components   = { workspace = true }
-hermes-runtime-components   = { workspace = true }
-hermes-logging-components   = { workspace = true }
+cgp                             = { workspace = true }
+hermes-relayer-components       = { workspace = true }
+hermes-chain-type-components    = { workspace = true }
+hermes-runtime-components       = { workspace = true }
+hermes-logging-components       = { workspace = true }
