+++ crates/relayer/relayer-components-extra/src/components/extra/closures/chain/message_sender.rs
@@ -1,4 +1,5 @@
 use cgp::prelude::HasComponents;
+use hermes_chain_type_components::traits::fields::height::CanIncrementHeight;
 use hermes_relayer_components::chain::traits::message_builders::update_client::{
     CanBuildUpdateClientMessage, UpdateClientMessageBuilder,
 };
@@ -23,7 +24,6 @@ use hermes_relayer_components::chain::traits::types::client_state::{
     HasClientStateFields, HasClientStateType,
 };
 use hermes_relayer_components::chain::traits::types::consensus_state::HasConsensusStateType;
-use hermes_relayer_components::chain::traits::types::height::CanIncrementHeight;
 use hermes_relayer_components::chain::traits::types::ibc::{
     HasCounterpartyMessageHeight, HasIbcChainTypes,
 };
